+++ packages/deployments/contracts/contracts/core/connext/facets/upgrade-initializers/DiamondInit.sol
@@ -34,7 +34,7 @@ contract DiamondInit is BaseConnextFacet {
   // You can add parameters to this function in order to pass in
   // data to set your own state variables
   function init(
-    uint256 _domain,
+    uint32 _domain,
     address _xAppConnectionManager,
     address _tokenRegistry, // Nomad token registry
     address _wrappedNative,
